Update dark theme script Added enabled extension condition to button
